Fix regexp filters to not match partial lines
We use a lot of regular expression to match filters, and we did not
include start and stop of line anchors. This caused the filters to match
everything that had a substring with the value we filtered for.

E.g. searching for site "napoli" would match site "indianapolis" since
"napoli" is a substring of "india-napoli-s".

This PR add anchors like "^napoli$" to only match if the start and end
of line also matches.