Add files for event support
This commit adds configuration and helper files for event support in the time-in-area example.
These files will monitor AOA detections and send a stateful event when a detection has been in the monitored zone for too long (expressed as active=1), as well as when the detection leaves the monitored area (expressed as active=0).